felix/bpf: bpfEndpointManager marks weps as ready.
When bpfEndpointManager has ifindex and applies tc programs to a WEP, it
also records in a bpf map that the iface is "ready". Thereafter tc
programs that have successful FIB lookup can redirec straight to the
wep. Otherwise the packet is dropped as it is unsafe to forward it and
it it pointless to let it be dropped by iptables for the same reason.