Extract manifest generation from the docs build (#6020)
Remove manifest generation from calico/ directory. Instead, add new charts/ directory and move all manifests to the top-level manifests/ directory, with a new script to re-generate manifests based on the charts.

This also includes changes to our OCP install instructions and docs build, as well as release process in order to ensure manifests are up-to-date as part of release.