Fix announcement of /32 and /128 entries for serviceLoadBalancerIPs.
After this patch, single IPs are no longer announced as global routes.

The patch also fixes a bug related to config change, causing one or more route to be withdrawn if changing from
for example /32 to /29 in BGPConfiguration. Now setRoutesForKey() assures that routes it thinks are advertised
actually do exist.