felix/bpf: map upgrades do not need caching map
The added value is little and it prevents the chaching map to be typed.
We can just upgrade the k-v pairs as we go. We cache them in memory like
the caching map would do because writing them back in the iterator would
not be able to return an error and break the iterator in a simple way.