felix/bpf: service routes are managed by a caching map
The caching map ensures that the desired state does not miss an update
when, for example, a service IP is reause. That is, a service gets
deleted and a new service with the same IP gets added in the same calc
graph iteration, etc.

The caching map updates the routetable.RouteTable which then ensures
that the desired state is maintained in the dataplane.