felix/bpf: service routes maintenance fixed
This fix ensures that if a service IP is recycled during the same calc
graph iteration, then we endup with the IP being present in the route
table.

Previously, it was possible to be programmed and then deleted by a service
that was being removed if iterating over the dirtyServices was done in a
wrong order.

We remove all IPs immediately when they are deleted, while we add IPs
of dirtyServices that should exist at the end of the iteration.