felix: FeatureGates - to dis/enable features
Sometimes we want to be able to include code that may not be finished yet
or 100% tested and we want to keep that code disabled in normal
production environments but we want to have it enabled in tests or in
specific user environments. Or we have a fix that breaks other things but
works for a particular corner case usecase without troubles.

In such cases we want to have the flexibility to configure felix easily
without introducing new config parameters that may disappear soon etc.

FeatureGates works similar to FeatureDetectOverride which we
previously overloaded for this functionality. However,
FeatureDetectOverride is specifically used to overriding environment
feature detection in cases where the automatic detection mechanism is
not perfect and configures felix in a troublesome way.