felix/bpf: pod-service-self should not rely on src==dst
Mark the connections if the post dnat dst is the same as src when
resolving on WEP and reuse that information for subsequent packets.

src==dst may be equal in other scenarious like host talking to a pod on
the same host via its own ip (nodeport). Than the communication is
host-host on one leg.