Increment Debian packaging epoch to "2:"
We use `git cherry -v ${last_tag} | wc -l` to count the number of commits since the last `0.dev`
development tag on master, and include that count in the Debian version so that it is always
increasing.

For some reason, though, that number is now NOT increasing!  The current packages in our master PPA
have a count of 295, but that command on the current monorepo gives 176.  Huh?

I don't really know why that is, but possibly connected with Semaphore using a shallow checkout
previously?  I've fixed that in #6763, but now we also need to bump the epoch number so that we can
start a new sequence of versions from 176.