In fact don't need to check for interface being up at all
As @fasaxc has pointed out, we don't actually need to check for interface state, because the
TAP interface will obviously be as up as it needs to be when a VM is actually using it.  Let's
simplify just to track whatever latest MTU we have seen, regardless of interface state.

The other important detail is to clean up the map storage when a TAP interface is deleted, but again
that is unrelated to the interface states that we have seen.