felix/fv: always test with UDP ctlb disabled
This is the desired production state, we want the UDP workaround instead
of the CTLB since we need to be able to react to changing UDP service
backend.

We test UDP both w/ and w/o CTLB eventhough we always use the workaround
because it also makes sure that the UDP services are programmed
(selectively) in routing table. When CTLB is completely disable all
services are programmed.