networking_calico: Prime project cache at startup
The Calico mechanism driver maintains a cache of project
data that it uses to annotate ports with. This cache is
currently filled whenever an item is found missing at
runtime.

This means, though, that the first port annotation
performed after a cold restart of `neutron-server` will
result in a fill penalty that could be avoided by
priming the cache during driver startup.

This commit thus primes the cache during startup.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>