networking_calico: Fix spurious endpoint delete
A port which is created and subsequently updated may race
into code which was meant to delete the WorkloadEndpoint on
the source hypervisor during a live-migration.

When the mechanism driver gets a port-update, first check
that the binding existed to begin with before deleting the
WorkloadEndpoint. If no such binding exists on the original
port, then the port just came into existence and was
updated just very shortly after.

When this happens, logs like the following are printed:
"Migration, delete WorkloadEndpoint on old host"
(rather than a message with an actual hostname to which
the original port was bound).

The fact that this was happening should have been harmless
and just resulted in some confusing log messages, but fix
it nonetheless.

Signed-off-by: Tyler Stachecki <tstachecki@bloomberg.net>