Retain OpenSSL FIPS dependent files in calico-node image
The current calico-node cleanup script removes certain files from the base
image that FIPS mode relies on. When FIPS is enabled, OpenSSL gives internal
error `FATAL FIPS SELFTEST FAILURE`. It also causes calico-node readiness
probe to fail with the same error and become not ready.

The checkup script is updated to not remove libssl.so and hidden hmac
files.