felix/bpf: mark CALI_ST_SKIP_FIB packets on ingress of heps
Disable FIB, let the packet go through the host after it is
policed. It is ingress into the system and we do not know what
exactly is the packet's destination. It may be a local VM or
something similar and we let the host to route it or dump it.

https://github.com/projectcalico/calico/issues/6450