felix/bpf: pod->svc->self does not need allow policy
We allow this traffic to match iptables dp and not require to have a
special policy that allows traffic based on its host IP which is hard to
express when pods start and go.

The check is done after we parsed the packet and we do it before policy
and before CT. So we only do it for the first packet of a connection
with minimal overhead.