Add flag to calico windows uninstall script and pass it from hostprocess install
The calico windows uninstall script would not run fully when it detects
it is running in a hostprocess container, but it is called by the
hostprocess installation to clean up a previous manual install, so
it needs to run fully in this special case. Fixing it by adding a flag
to indicate this situation.