typha: Fix missing nsswitch files causing localhost lookup fails
Between the calico/typha:v3.23.5 and calico/typha:v3.24.0 Docker images,
the included C libraries were reduced from many to only a handful.
However, most prominently, it lacked libnss_files, which is needed to do
proper lookups from /etc/hosts.

In some environments, things still worked. In others, the 'localhost'
lookup would turn into a 'localhost.the-domain-the-pod-is-in' lookup,
which could become something unexpected. For instance when '*' CNAMEs
are used.

The result of the 'localhost' lookup is used to decide on which address
the health check responder should bind. For 127.0.0.1, it would bind
just fine. For an external IP, it would fail:

  health.go 296: Health endpoint failed, trying to restart it...
    error=listen tcp REMOTE_IP:9098: bind: cannot assign requested address

This in turn caused the pod to never become Ready.

Interestingly, the 'options' parameter in /etc/resolv.conf also affected
the behaviour. Only when *any* 'options' was set AND the 'search'
parameter was set, would this result in such a faulty lookup.

This changeset:

- removes the duplicate setting of nsswitch.conf
- sets only nsswitch libraries that are actually available

An alternative fix would be to build typha with CGO_ENABLED=0. Golang
has builtin support for both 'dns' and 'files' lookups, so there is no
reason to use the nsswitch subsystem at all.

Fixes: #6710