bpf/ut: TestNATAffinity flakes because of a too short timeout
The test may take longer than the timeout,mainly because of loading the
test bpf programs.

The length of the timeout is not important for the test because the test
expires the NAT selections artificially to make the test deterministic
and without a need for the timeout to trully expire.

Co-Authored-By: Sridhar Mahadevan <sridhar@tigera.io>