[BPF] detach cali programs from devices no longer in regex
It can happen that the device regexp changes and calico programs then
remain attached to devices that we do not observe anymore. For instance,
user has the default ethX in the regexp, but those devices are bonded
and the user does not have bondX in the regexp. The user fixes the
regexp to include bondX only, but then the programs stay attached to
ethX. User has to remove the programs manually or restart the node.

We figure out that the programs on devices not in regexp are ours when
those programs have a jump map pinned inthe calico bpffs hierarchy.

This is not really a problem for workload devices as they usually
disapear and programs get detached then.