Optimise KeyFromDefaultPath.
The old version was linear, trying one key type after another.
As the number of key types has grown this now causes noticable CPU
usage when benchmarking. This commit reworks the algorithm to
parse keys using a tree of switch statements and other conditions.

I considered making this a bit more generic but there are a lot of
exceptions and special cases (for example, Felix config keys allow
slashes in the name of the key), and, this should be a one-off since
all new resoruces will go through the "resources/v3" path.

Add a Fuzz test that compares with the old implementation.