[BPF] avoid calling bpftool and tc to check if things changed
Collect information on all attached programs at the top of
CompleteDeferredWork and remember that for the duration of it. That is a
single call to bpftool if there is any change instead of calling and
parsing it for each modified device.

We also called the bpftool and tc tool multiple times when things
changed, especially when felix restarted.