[BPF] no need for CALI_SKB_MARK_BYPASS_FWD_SRC_FIXUP mark
We do not have a usecase for fixing up host IP when a packet is leaving
the host. We always set the source when we do vxlan encap as expected.
Therefore we do not need to ever set the mark and check for it, which
makes the code simpler a slightly faster.