[BPF] separate objects for v4/6 programs
We do not need to load programs that we are not going to use. When
needed, the v6 object is loaded separately and the programs get included
in the jump map.

FV tests no longer load v6 programs, but the programs are tested with
ap.AttachProgram in UTs. FVs should be a little faster to update
endpoints. Nothing changes for production except that the code is never
loaded into kernel - so a little faster too.

UTs do not excercise TestIPv6Parsing as that is a little bit
pointless and fixing how the programs would load in UTs for BPF_RUN is
only worth exploring with further progress on v6.