OpenStack: Support newer, fixed version of etcd server
etcd prior to v3.4.20 had a scalability bug that limited the number of possible simultaneous client
watches, and hence the maximum number of cluster nodes in a Calico+OpenStack installation.  Hence
we'd like to support etcd v3.4.20 and later.

However, etcd v3.4 only supports URL paths with '/v3beta/' or '/v3/', and not the '/v3alpha/' that
is hardcoded in our documented version of the etcd3gw package (previously known as etcd3-gateway).
Hence we also need to:

- document using a newer etcd3gw that supports the newer URL paths, namely version 1.0.1

- add a Calico config option to control that (bearing in mind that etcd3gw's default is still
  '/v3alpha/', and that some users might still be using older etcd servers, and so we can't just
  hardcode a newer setting).