Fix bug with port status checking in test_snapshot
We had a bug in product code that was recently fixed by
https://github.com/projectcalico/calico/pull/6979.  That bug caused some port statuses to be
signalled as "None" when they should not be.

Unfortunately we also had this test (test_snapshot) that had been coded to match the bug, i.e. to
expect a port status update of "None".  The correct behaviour is that the port status should not be
changed by this passage of test code.

This was missed at the time of the 6979 PR because our UT running had a bug that allowed some tests
to fail but still reported success overall.