[BPF] BPFDSROptoutCIDRs can overlaps with other CIDRs
For instance 0.0.0.0/0 would optout all clients from DSR. Those CIDRs
must be either disjoint with other CIDRs in the cluster or their
super-sets. That means, for instance, if an IP pool overlaps only
partially with the opt out CIDR, the optout may or may not be applied as
it is applied to the route blocks on each node, but does not split those
blocks.

The BPFDSROptoutCIDRs is meant to be used with external entities to the
cluser. It has no effect on in-cluster entities since DSR does not
have much meaning fot them anyway.

A useful case may be to apply this to external nodes, provider specific
ip like AKS healt check IPs etc.