Avoid logging WARNINGs about liveness/readiness during initial startup
With Skimble we sometimes see WARNINGs about non-readiness during the first 10 seconds or so of
operation.  This is just because Skimble has not yet reached its inital ready state.

Those WARNINGs are benign, so we'd like to suppress them for Skimble - and more generally it makes
sense that what really matters is regression from a previously live or ready state.  So tweak the
coding so as only to Warn after previously being live or ready.