Let UID be stored and retrieved by annotation instead of cloning UIDs.
When converting from calico to k8s objects and back, there is already logic storing the ObjectMeta in an annotation for the conversion. There should be no need to set the UID of the converted object to match the source object, and it's a bad practice because kubernetes expectes UIDs to be unique from one object to the next for operations like garbage collection.