[BPF] install log filter programs based on BPFLogFilters option
If BPFLogFilters is set only specified packets on specified devices
trigger log actions. In any other case packets go through without the
overhead of logging.

A filter program is called by the preamble and that program decides
whether packets should be processed by a set of programs with or without
logging. This effectively means that there are two paths that a packet
can take when logging is enabled.

If BPFLogFilters is not specified, that filter program matches all
packets and every program logs for every packet like the current
firehose.

As CTLB logging also goes to the same trace pipe, it mixes with tc logs
and thus CLTB logging must be explicitly enabled if BPFLogFilters is
specified by including 'ctlb=on' to avoid logging more than desired.
CTLB is still a firehose.