Add deduplicating syncer callbacks buffer.
The deduplicating buffer implements the syncer callbacks
API on its input, and calls another syncer callback on its output.
In-between it maintains an in-order queue of KV updates and a
tracking map to record what is in the queue.

If an update comes in for a key that already has a value in the queue
then the value in the queue is replaced with the updated KV.

This can cause reordering (which is allowed by the syncer API)
but it ensures that the amount of KVs in flight is bounded to the
total size of the datastore even under substantial overload.
The effect is that the client will periodically "skip ahead"
to the more recent state of the datatstore without seeing
intermediate states.