Fix docker image folder permissions
Under some circumstances, the apiserver can and
should be run as non-root. Previously, the docker
image did not permit this because folders /code
and /tmp were solely writable by root.

In the existing Dockerfile, it's indicated that
both directories should be writable for user 1001.
However, this logic is flawed due to:

1) The UBI running as root, resulting in permissions
   after mkdir at drwxr-xr-x owned by root:root
2) The COPY function not honoring folder permissions,
   even with --chmod, which again leads to
   drwxr-xr-x owned by root:root

To resolve this, we employed the fact that ADD
extracts .tar files while preserving the permissions.
Now, the file structure is:

/code drwxrwxrwx
|-- apiserver -rwxr-xr-x
|-- filecheck -rwxr-xr-x
/tmp drwxrwxrwx