[BPF] Do not allow forwarding icmpv6 bootstrap messages
BPF program on tc hooks allow the messages in and out, but they must be
allow only between endpoints and the host. No forwarding must be
allowed. We always force the messages into iptables on any ingress to
the host and we make iptable rules to make sure that they are dropped if
forwarded.