[BPF] icmp errors must be always treated as related
There is no reason why error would be generated if there was not already
a related data stream. We should not treat icmp as related if there is
no CT match yet. In ICMPv6 there is a high probability (it is pretty much
certain that there is alredy a CT entry for ICMPv6 stream because of
usage for ipv6 bootstrap, neighbor and router lookup etc. Related
traffic should not match on such CT entries and passed unmodified in
case of NAT etc.