[BPF] fix ct cleaning for host processes w/o ctlb using service
If there is an overlay, forwarding entry may not have a matching client
IP as it may start with host's main IP, but then it is turned into
host's tunnel IP. We cannot determine what is the client/service/backend
from that entry only. We need to lookup the reverse entry. If the
reverse entry does not exist, we delete the forarding entry anyway as
there is no use for it without the reverse half.