Patch calico-felix libpcap soname for Openstack
This changeset patches calico-felix binary libpcap soname for Debian
and its derived OSes. We switched to RHEL/UBI based go-build system
which links to a different libpcap soname. The actual libpcap ships with
Debian is at 1.x but the soname is still at 0.8. calico-felix binary is
patched during packaging to the older soname. This change also bumps
compat version to v10 that is recommended in [1] for new packages.

[1] https://www.debian.org/doc/manuals/maint-guide/dother.en.html#compat