[BPF] when host IP change kube-proxy replaces only Syncer
Restarting the whole proxy is an overkill, only the syncer eed the set
of host IPs. Also the health check servers created by kube-proxy
frontend are not easily carried over and thus restarting kube-proxy
cannot restart the health checks.

As a side-effect, the "restart" logic in kube-proxy is much simpler as
we do no tneed to stop that much.