pods in nat-outgoing shoul dnot SNAT to local host
When a pod is accessing a local host, it should not get SNATed as the
host when it is in a nat-outgoing ippool. (a) it is unnecessary as the
local node can be accessed and (b) there is no way to return the traffic
as is it would return to the host itself.

refs https://github.com/projectcalico/calico/issues/7252