suppress normal HEP policy on WEP if there is no wildcard HostEndpoint
If there is no wildcard HEP, there is no policy that should be applied.
But without skipping, empty list of profiles would create a default deny
rule if none of the non-existent profiles matches. That is obviously
always hit and traffic toward the host is dropped if
defaultEndpointToHostAction is set to RETURN.

fixes https://github.com/projectcalico/calico/issues/7252