[CORE-10110] Fix confd on Windows HPC
Use winutils.GetHostPath() for file paths on confd and typha so that
they work correctly on windows HPC (typha runs on linux only, but
confd reuses the env vars and code that deals with cert mounts).

Add necessary confd files to cnx-node-windows image so that the
confd container functions correctly.

Add the correct dir for powershell.exe to $PATH so that confd
can use it properly on windows HPC.

Only add powershell path to $PATH if it can't be found in the first place.