[BPF] proxy - syncer must not count not-ready endpoints
We include terminating eps in the eps list so that we can prevent
conntrack clean up, but they are not written into the backend map and so
they should not be reflected in the backends count.

Test that Terminating is propagated through proxy frontend.
Test in FV that proxy removes the terminating backend from the NAT
table.

refs https://github.com/projectcalico/calico/issues/8423