Adjust creation of same-subnet routing table to include previous parent names.
This only fixes the original issue as long as felix stays running
while the parent device changes.  However, it doesn't have the
unintended consequences of always removing conflicting routes.

Namely: if we always remove conflicting routes, then two of our
RouteTables can fight over a route without hitting any errors.
This means that, when the conflict is resolved in the datamodel,
the RouteTables won't "fix up" because they don't know anything
is wrong.