[BPF] use label to clean up conntrack to terminating udp backends
When a backend pod of a service is in terminating state, we keep existing
connections as that pod is likely in a terminating grace period.
However, a pod may get stuck in terminating state. New connections are
not directed to that pod and if there is no connectivity to the pod TCP
will fail eventually and new connections will select another backend.
However UDP follows the established conntrack. So labeling a service
with

projectcalico.org/epbf-udp-remove-terminating-conntrack = true

would ensure that we clean up conntrack entries towards this backend
once it becomes terminating. This make the dataplane pick another
backend once there is a conntrack miss, which unblocks the client.

Fixes https://github.com/projectcalico/calico/issues/8423