[BPF] police MASQed packets from pod->service->self as if from pod
To match the behaviour of iptables, we need to do the policy on the
original source (self) instead of on the hosts source after MASQ so that
policy applies correctly. That is, if the policy says the ingress is
allowed from any pod matching a selector and if self matches the
selector, the packet must be accepted. Matching after MASQ breaks that.