[BPF] Update syncer status from dataplane later
When swapping a syncer in kube-proxy, defer loading state from dataplane
until after the old syncer is done. Otherwise we may load state that is
still being updated. This could lead to a discrepancy between the real
state of the dataplane and what the caching maps thing the real state is.