host can access self via a service without CTLB
added test for this case

No host source conflict resolution on lo since we cannot tell traffic
straight to self from traffic to self via a service. Host connecting to
self and connecting to the same self port via a service should be rare,
but if that is necessaty, than CTLB is the only option.