Fix calico/apiserver /tmp folder permission
This change backports changes from [1] for `calico/apiserver` only to
fix the `/tmp` folder permission. Currently, when `/tmp` folder is
copied form ubi stage to scratch, the folder permission is 0755 so it isn't
writable for non-root processes. With the changes in [2], apiserver won't
be ready because `/tmp/ready` flag can't be written.

[1] https://github.com/projectcalico/calico/pull/8299
[2] https://github.com/tigera/operator/pull/2906