[BPF] When lo has IP override it when CTLB is disabled
When CTLB is disabled, we route traffic for services via bpfnatin/out
device. Since the final destination isn't resolved yet, Linux picks up
an address set on loopback device (if there is any) as source. This may
not be (likely is not) an address that can be used by the destination to
return traffic. Therefore we need to replace it by the host's IP that is
routable within the cluster. So we set the host IP as a source on the
service routes.

We can only set the routes when there is a host IP, but that is also a
prerequisite for loading any bpf programs. When/if the host IP changes,
we mark all services as dirty so we reapply them with updated source.