[BPF] When lo has IP override it when CTLB is disabled
When CTLB is disabled, we route traffic for services via bpfnatin/out
device. Since the final destination isn't resolved yet, Linux picks up
an address set on loopback device (if there is any) as source. This may
not be (likely is not) an address that can be used by the destination to
return traffic. Therefore we need to replace it by the host's IP that is
routable within the cluster.

We use the same mechanism as for replacing main host device IP with a
tunnel IP when we need to reach a remote workload via an overlay.