Add logging statement to indicate ss command usage
Added a fmt.Println statement to log a message explaining that the  
command "netstat -a -n" failed to run and that "ss -a -n" was used  
instead. This change was made to provide better context to users  
who might encounter the error message "Failed to run command: netstat -a -n".  
  
Users who encounter this error should not immediately consider installing netstat.