When IPv6 is enabled, either v4 or v6 connectivity is enough
If IPv6 is enabled (which enables dual stack), it is ok, if only ipv6
addresses are avilable. Felix reports ready and connectivity exists.
When IPv4 are assigned later, IPv4 connectivity will work. The same hold
in reverse order.

worded differently, Felix reports ready if the path for whatever ip
version of its host IPs is ready. If there is a vX host IP and version
X path is not loaded, Felix is not ready as there is an error.