[BPF] Variable source port for VXLAN NodePort tunnel
To efficiently use bonded network cards or multiple CPUs on the
receiving node, it is wrong to use a single VXLAN stream for forwarding
services (node ports) from one node to a backend on another node. We
need to preserve the notion of separate flows by varying the source port
for each of the flowsi - as recommended by the RFC. We do so be XORing
the source ports of the original stream to create the source port for
the vxlan stream that carries it.

refs https://github.com/projectcalico/calico/issues/8620