Add option to disable announcement of aggregated LoadBalancerIP ranges
This MR allows to disable the default announcement of all
LoadBalancerIP ranges by setting `serviceLoadBalancerRouteAggregationEnabled: false` in the BGPConfiguration.

We are advertising IPs of the same `/24` in multiple clusters. Creating
LoadBalancer services with externalTrafficPolicy Local are reachable
because additional /32 route are being announced only from the cluster
where the LoadBalancer exists. But LoadBalancers with
externalTrafficPolicy Cluster are not working since the /24 route is
being announced from multiple clusters and traffic not reaching the
correct cluster.
To mitigate this we would have to add every ip as a /32 to the
BGPConfiguration.

This MR adds an additional option to the BGPConfiguration to disable
the default annoncement of all prefixes in the `serviceLoadBalancerIPs`
list. In which case we are also announcing /32 routes for LoadBalancers
with externalTrafficPolicy Cluster.