send RST to midflow TCp packets from unknown streams
For a smooth switch from iptables to ebpf mode, we do not want to
interrupt existing connections. If we see midflow packets, we pass them
to the host stack. If the stack can verify that they belong to an
existing conntrack, we let them through and we learn the conntrack.

We drop the rest. However, there are some situations when we can see a
stray TCP packet during ebpf mode, for instance when a pod dies and ECMP
kicks in and sends a packet to a different host.

If such a packet gets dropped, the end of the connections remains stuck.
This change sends an RST to such a stream instead of just dropping the
packets so that the end host can break the connection.

Fixes https://github.com/projectcalico/calico/issues/8882