[BPF] set MTU on bpfin/out.cali to match host
Set the MTU to match the smallest MTU of the host devices as per the
current autodetection. That makes sure that if large MTU is used, the
extra device does not create a bottleneck with a small MTU. Also when
MTU changes, on host devices, it get adjusted automatically. If an
overlay is used, the special device MTU is adjusted to the size of the
overlay.

Fixes https://github.com/projectcalico/calico/issues/8918