Fix libpcap patchelf for calico-felix
For Calico v3.27, we build felix with Debian-based toolchain which
links against with libpcap.so.0.8. It will fail on RHEL rpm install because
RHEL ships with libpcap.so.1. This change patchelf calico-felix to run
on RHEL distros. This is not required for Calico v3.28+ because we
switched our building base to RHEL/UBI.