Rebuild flannel-cni-plugin binaries on demand
Do the same for flannel-cni-plugin as https://github.com/projectcalico/calico/pull/8909
did for containernetworking-plugins, i.e. clone the repo and build using
the go-build image from the calico monorepo on-demand.

This is done to remove the tedious step of creating new releases and updating
the Makefile with the new version on the flannel-cni-plugin repo.