Execute MASQ rules after kube-proxy to avoid conflict
Calico needs to do MASQ for several reasons and kube-proxy does it when
it forwards nodeports to another node. Calico conflicts with this MASQ
in case it routes traffic to the destionation via tunnel. This may
result in a conflict which triggers wrongly calculated checksums in case
tx csum offloading is enabled on the tunnel device, which lead to
disabling offloading on the tunnel and thus to significant performance
hit.

Do MASQ after kube-proxy means that if kube-proxy already did it,
calico's is not executed - it would be a duplicate anyway. Other MASQ
and SNAT cases in the calico chain are othogonal to the kube-proxy
usecase and thus execute as kube-proxy does not do MASQ in that case.

Note that unlik ein other chains where calico may do policy and thus
needs to go first, calico does not execute any policy in POSTROUTING
nat.